Move EvolvedOperatorAnsatz to NLocal directory (#11385)
* Move EvolvedOperatorAnsatz

* add release note fix import

* fix import

* add doc string to function

* fix imports

* Add tests for Moved Class

* update library init

* Fix library init imports

* Add deprecation to old class

* update formatting

* Fix tests

* fix import

* Move EvolvedOperatorAnsatz

* add release note fix import

* fix import

* add doc string to function

* fix imports

* Add tests for Moved Class

* update library init

* Fix library init imports

* Add deprecation to old class

* update formatting

* Fix tests

* fix import

* Change deprecation to upgrade

* Fix init doc

---------
