+++ qiskit/circuit/library/n_local/__init__.py
@@ -17,6 +17,7 @@ from .two_local import TwoLocal
 from .pauli_two_design import PauliTwoDesign
 from .real_amplitudes import RealAmplitudes
 from .efficient_su2 import EfficientSU2
+from .evolved_operator_ansatz import EvolvedOperatorAnsatz
 from .excitation_preserving import ExcitationPreserving
 from .qaoa_ansatz import QAOAAnsatz
 
@@ -26,6 +27,7 @@ __all__ = [
     "RealAmplitudes",
     "PauliTwoDesign",
     "EfficientSU2",
+    "EvolvedOperatorAnsatz",
     "ExcitationPreserving",
     "QAOAAnsatz",
 ]
