+++ qiskit/circuit/library/n_local/evolved_operator_ansatz.py
@@ -17,14 +17,14 @@ from collections.abc import Sequence
 
 import numpy as np
 
+from qiskit.circuit.library.pauli_evolution import PauliEvolutionGate
 from qiskit.circuit.parameter import Parameter
 from qiskit.circuit.quantumregister import QuantumRegister
 from qiskit.circuit.quantumcircuit import QuantumCircuit
 from qiskit.quantum_info import Operator, Pauli, SparsePauliOp
 from qiskit.synthesis.evolution import LieTrotter
 
-from .pauli_evolution import PauliEvolutionGate
-from .n_local.n_local import NLocal
+from .n_local import NLocal
 
 
 class EvolvedOperatorAnsatz(NLocal):
@@ -98,6 +98,11 @@ class EvolvedOperatorAnsatz(NLocal):
 
     @property
     def num_qubits(self) -> int:
+        """Returns the number of qubits in this circuit.
+
+        Returns:
+            The number of qubits.
+        """
         if self.operators is None:
             return 0
