+++ releasenotes/notes/move_evolved_operator_ansatz-5406bc071013c443.yaml
@@ -1,0 +1,9 @@
+---
+upgrade:
+  - |
+    The class :class:`.EvolvedOperatorAnsatz` has been migrated from
+    :mod:`qiskit.circuit.library` to :mod:`qiskit.circuit.library.n_local` to match
+    the location of its base class. This move has not affected the usual import path
+    of the class: ``from qiskit.circuit.library import EvolvedOperatorAnsatz``, but
+    now it is also accessible doing ``from qiskit.circuit.library.n_local import
+    EvolvedOperatorAnsatz``.
