+++ test/python/circuit/library/test_evolved_op_ansatz.py
@@ -17,7 +17,8 @@ import numpy as np
 from qiskit.circuit import QuantumCircuit
 from qiskit.quantum_info import SparsePauliOp, Operator, Pauli
 
-from qiskit.circuit.library import EvolvedOperatorAnsatz, HamiltonianGate
+from qiskit.circuit.library import HamiltonianGate
+from qiskit.circuit.library.n_local import EvolvedOperatorAnsatz
 from qiskit.synthesis.evolution import MatrixExponential
 from qiskit.test import QiskitTestCase
 
@@ -51,11 +52,11 @@ class TestEvolvedOperatorAnsatz(QiskitTestCase):
         reference.append(HamiltonianGate(matrix, parameters[0]), [0, 1, 2])
 
         decomposed = evo.decompose().decompose()
+
         self.assertEqual(decomposed, reference)
 
     def test_changing_operators(self):
         """Test rebuilding after the operators changed."""
-
         ops = [Pauli("X"), Pauli("Y"), Pauli("Z")]
         evo = EvolvedOperatorAnsatz(ops)
         evo.operators = [Pauli("X"), Pauli("Y")]
@@ -79,6 +80,7 @@ class TestEvolvedOperatorAnsatz(QiskitTestCase):
         for parameter in evo.parameters:
             ref.rz(2.0 * parameter, 0)
             ref.barrier()
+
         self.assertEqual(evo.decompose(reps=2), ref)
 
     def test_empty_build_fails(self):
